[refactor] relocate modal files
