+++ src/features/modal/components/index.ts
@@ -1,6 +1,3 @@
-export { default as ModalContainer } from "./modalContainer";
-export { default as Modal } from "./modal";
-
 export { default as BasicModalHeader } from "./header/basicModalHeader";
 export { default as NumberedModalHeader } from "./header/numberedModalHeader";
 export { default as BasicModalFooter } from "./footer/basicModalFooter";
