+++ src/features/modal/components/modal.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 interface ModalProps {
   isOpen: boolean;
   onClose: () => void;
