+++ src/features/modal/components/tab/modalTab.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { ReactNode, useState } from "react";
 
 interface TabItem {
@@ -20,7 +18,6 @@ export default function ModalTab({ tabs, defaultTab }: ModalTabProps) {
 
   return (
     <div className="flex flex-col flex-1 overflow-hidden">
-      {/* 탭 헤더 */}
       <div className="flex border-b border-gray-200 flex-shrink-0">
         {tabs.map((tab) => (
           <button
@@ -37,7 +34,6 @@ export default function ModalTab({ tabs, defaultTab }: ModalTabProps) {
         ))}
       </div>
 
-      {/* 탭 컨텐츠 - 스크롤 영역 */}
       <div className="flex-1 overflow-y-auto p-6 scrollbar-custom">
         {activeTabContent}
       </div>
