+++ src/features/modal/index.ts
@@ -1,2 +1,3 @@
-export { useModalStore } from "./modalStore";
+export { default as ModalContainer } from "./containers/modalContainer";
+export { useModalStore } from "./stores/modalStore";
 export * from "./components";
